Update examples and README.md to make `usage` easier to understand
