+++ examples/example_phrase/dict.txt
@@ -1,0 +1,1 @@
+都会区 100 n
